requirements.txt
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
cached-property==1.5.2
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.7
django-six==1.0.4
django-uuslug==1.2.0
filelock==3.0.12
idna==3.2
isodate==0.6.0
lxml==4.6.3
mysqlclient==2.0.3
Pillow==8.3.2
python-slugify==5.0.2
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
slugify==0.0.1
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
virtualenv==20.7.2
xlwt==1.3.0
zeep==4.1.0
